Tooltips almost perfect
I've removed them from edge, made them look the same in chrome and firefox, and they are mostly good. Only 3 minor problems remain.  Lines 188-192 of theme.js only work the first time they are run. When an svg path is clicked the event.target is the path, not the svg itself. And the arrows are sometimes not fully connected to the body of the tooltip. Oh, and the menu is still completely broken in edge